requirements.txt
Flask==1.1.1
functions-framework==1.0.0
pyyaml==5.3
requests==2.22.0